+++ app.py
@@ -838,6 +838,7 @@ from app_clap_search import clap_search_bp
 from app_mulan_search import mulan_search_bp
 from app_backup import backup_bp
 from app_dashboard import dashboard_bp
+from app_users import users_bp
 
 app.register_blueprint(chat_bp, url_prefix='/chat')
 app.register_blueprint(clustering_bp)
@@ -857,6 +858,7 @@ app.register_blueprint(mulan_search_bp)
 app.register_blueprint(backup_bp)
 app.register_blueprint(migration_bp)
 app.register_blueprint(dashboard_bp)
+app.register_blueprint(users_bp)
 
 # --- Startup: Load indexes and caches (Flask server only, NOT RQ workers) ---
 # RQ workers import app.py but should NOT load indexes or start background threads.
