+++ templates/sidebar_navi.html
@@ -1,6 +1,8 @@
 {% if setup_saved %}
 <li><a href="{{ url_for('dashboard_bp.dashboard_page') }}" {% if active == 'dashboard' %}class="active"{% endif %}>Dashboard</a></li>
+{% if is_admin %}
 <li><a href="{{ url_for('index') }}" {% if active == 'index' %}class="active"{% endif %}>Analysis and Clustering</a></li>
+{% endif %}
 <li><a href="{{ url_for('chat_bp.chat_home') }}" {% if active == 'chat' %}class="active"{% endif %}>Instant Playlist</a></li>
 <li><a href="{{ url_for('voyager_bp.similarity_page') }}" {% if active == 'similarity' %}class="active"{% endif %}>Playlist from Similar Song</a></li>
 <li><a href="{{ url_for('artist_similarity_bp.artist_similarity_page') }}" {% if active == 'artist_similarity' %}class="active"{% endif %}>Artist Similarity</a></li>
@@ -13,13 +15,14 @@
 <li><a href="{{ url_for('waveform_bp.waveform_page') }}" {% if active == 'waveform' %}class="active"{% endif %}>Waveform</a></li>
 <!-- <li><a href="{{ url_for('collection_bp.collection_page') }}" {% if active == 'collection' %}class="active"{% endif %}">Collection Sync</a></li> -->
 {% if is_admin %}
-<li class="has-submenu{% if active in ('cleaning', 'cron', 'backup', 'provider_migration', 'setup') %} open{% endif %}">
+<li class="has-submenu{% if active in ('cleaning', 'cron', 'backup', 'provider_migration', 'setup', 'users') %} open{% endif %}">
     <a href="#" class="submenu-toggle">Administration <span class="submenu-arrow">›</span></a>
     <ul class="submenu">
         <li><a href="{{ url_for('analysis_bp.cleaning_page') }}" {% if active == 'cleaning' %}class="active"{% endif %}>Cleaning</a></li>
         <li><a href="{{ url_for('cron_bp.cron_page') }}" {% if active == 'cron' %}class="active"{% endif %}>Scheduled Tasks</a></li>
         <li><a href="{{ url_for('backup_bp.backup_page') }}" {% if active == 'backup' %}class="active"{% endif %}>Backup and Restore</a></li>
 		    <li><a href="{{ url_for('migration_bp.provider_migration_page') }}" {% if active == 'provider_migration' %}class="active"{% endif %}>Provider Migration</a></li>
+        <li><a href="{{ url_for('users_bp.users_page') }}" {% if active == 'users' %}class="active"{% endif %}>Users</a></li>
         <li><a href="{{ url_for('setup_page') }}" {% if active == 'setup' %}class="active"{% endif %}>Setup Wizard</a></li>
     </ul>
 </li>
